AP_NavEKF: Prevent transients failing GPS check after landing
If the GPS is not glitching when we land, then we assume the GPS is good and wait 5 seconds until a fail can be declared to allow touchdown transients to clear.
This opens up a small window of vulnerability wrt GPS glitching but is acceptable due to the low probability.